Moved content to create_routes file
