+++ src/delivery_algorithm.h
@@ -1,7 +1,2 @@
 #include "package_distribution.h"
 
-void add_edge(graph_t *graph, int node_src, int node_dest);
-
-void display_matrix(graph_t *graph);
-
-void free_matrix(graph_t *graph);
